requirements.txt
torch>=2.0.0
torchvision>=0.15.0
accelerate>=0.20.0
diffusers>=0.18.0
transformers>=4.30.0
tqdm>=4.65.0
Pillow>=9.5.0